+++ vault_cli/cli.py
@@ -3,7 +3,17 @@ import logging
 import os
 import pathlib
 import sys
-from typing import Any, Dict, Mapping, NoReturn, Optional, Sequence, TextIO
+from typing import (
+    Any,
+    Dict,
+    Generator,
+    Mapping,
+    NoReturn,
+    Optional,
+    Sequence,
+    TextIO,
+    Tuple,
+)
 
 import click
 import yaml
@@ -220,12 +230,13 @@ def get_all(client_obj: client.VaultClientBase, path: Sequence[str], flat: bool)
     ),
 )
 @click.argument("name")
+@click.argument("key", required=False)
 @handle_errors()
-def get(client_obj: client.VaultClientBase, text: bool, name: str):
+def get(client_obj: client.VaultClientBase, text: bool, key: Optional[str], name: str):
     """
     Return a single secret value.
     """
-    secret = client_obj.get_secret(path=name)
+    secret = client_obj.get_secret(path=name, key=key)
     force_yaml = isinstance(secret, list) or isinstance(secret, dict)
     if text and not force_yaml:
         if secret is None:
@@ -238,16 +249,43 @@ def get(client_obj: client.VaultClientBase, text: bool, name: str):
     )
 
 
+def build_kv(attributes: Sequence[str]) -> Generator[Tuple[str, str], None, None]:
+    """
+    Converts a list of "key=value" to tuples (key, value).
+    If the value is "-" then reads the secret from stdin.
+    """
+    for item in attributes:
+        try:
+            k, v = item.split("=", 1)
+        except ValueError:
+            raise click.UsageError(
+                f"Expecting 'key=value' arguments. '{ item }' provided."
+            )
+        if v == "-":
+            v = click.get_text_stream("stdin").read()
+        yield k, v
+
+
 @cli.command("set")
 @click.pass_obj
-@click.option("--yaml", "format_yaml", is_flag=True)
 @click.option(
+    "--update/--clear",
+    default=True,
+    help="Update the current kv mapping or replace the its content",
+)
+@click.option(
+    "-p",
     "--prompt",
     is_flag=True,
-    default=False,
-    help="Prompt user for value using a hidden input.",
+    help="Prompt user for values using a hidden input. Keys name are passed as arguments",
 )
-@click.option("--stdin/--no-stdin", default=False)
+@click.option(
+    "--file",
+    "yaml_file",
+    default=None,
+    help="Read key/value mapping from a file. A filename of '-' reads the standard input",
+    type=click.File(),
+)
 @click.option(
     "--force/--no-force",
     "-f",
@@ -256,62 +294,49 @@ def get(client_obj: client.VaultClientBase, text: bool, name: str):
     help="In case the path already holds a secret, allow overwriting it "
     "(this is necessary only if --safe-write is set).",
 )
-@click.option(
-    "--strip/--no-strip",
-    default=True,
-    help="If --no-strip is set, value will not be stripped of surrounding whitespaces",
-)
-@click.argument("name")
-@click.argument("value", nargs=-1)
+@click.argument("path")
+@click.argument("attributes", nargs=-1, metavar="[key=value...]")
 @handle_errors()
 def set_(
     client_obj: client.VaultClientBase,
-    format_yaml: bool,
+    update: bool,
     prompt: bool,
-    stdin: bool,
-    strip: bool,
-    name: str,
-    value: Sequence[str],
+    yaml_file: TextIO,
+    path: str,
+    attributes: Sequence[str],
     force: Optional[bool],
 ):
     """
-    Set a single secret to the given value(s).
+    Set a secret.
 
-    Value can be either passed as argument (several arguments will be
-    interpreted as a list) or via stdin with the --stdin flag.
+    \b
+    You can give secrets in 3 different ways:
+    - Usage: vault set [OPTIONS] PATH [key=value...]
+      directly in the arguments. A value of "-" means that value will be read from the standard input
+    - Usage: vault set [OPTIONS] PATH --prompt [key...]
+      prompt user for a values using hidden input
+    - Usage: vault set [OPTIONS] PATH --file=/path/to/file
+      using a json/yaml file
     """
-    if prompt + bool(value) + stdin > 1:
+    if bool(attributes) + bool(yaml_file) > 1:
         raise click.UsageError(
-            "Conflicting input methods: use only one of --prompt, --stdin and "
-            "positional argument"
+            "Conflicting input methods: you can't mix --file and positional argument"
         )
 
     json_value: types.JSONValue
-    if stdin:
-        json_value = click.get_text_stream("stdin").read()
-
+    if yaml_file:
+        json_value = yaml.safe_load(yaml_file)
     elif prompt:
-        json_value = click.prompt(f"Please enter value for `{name}`", hide_input=True)
-
-    elif len(value) == 1:
-        json_value = value[0]
-
-    else:
-        json_value = list(value)
-
-    if format_yaml:
-        if not isinstance(json_value, str):
-            raise click.UsageError(
-                "Cannot pass several values when using yaml format. "
-                "Please use a yaml list instead."
+        json_value = {}
+        for key in attributes:
+            json_value[key] = click.prompt(
+                f"Please enter a value for key `{key}` of `{path}`", hide_input=True
             )
-        json_value = yaml.safe_load(json_value)
-
-    if strip and isinstance(json_value, str):
-        json_value = fix_whitespaces(string=json_value)
+    else:
+        json_value = dict(build_kv(attributes))
 
     try:
-        client_obj.set_secret(path=name, value=json_value, force=force)
+        client_obj.set_secret(path=path, value=json_value, force=force, update=update)
     except exceptions.VaultOverwriteSecretError as exc:
         raise click.ClickException(
             f"Secret already exists at {exc.path}. Use -f to force overwriting."
@@ -321,27 +346,16 @@ def set_(
     click.echo("Done")
 
 
-def fix_whitespaces(string: str) -> str:
-    """
-    Single line secrets are stripped of all surrounding whitespace, which most probably
-    got here by accident. Multiline secrets are stripped, but a single trailing new line
-    is enforced.
-    """
-    stripped = string.strip()
-    if "\n" in stripped:
-        stripped += "\n"
-    return stripped
-
-
 @cli.command()
 @click.pass_obj
 @click.argument("name")
+@click.argument("key", required=False)
 @handle_errors()
-def delete(client_obj: client.VaultClientBase, name: str) -> None:
+def delete(client_obj: client.VaultClientBase, name: str, key: Optional[str]) -> None:
     """
     Delete a single secret.
     """
-    client_obj.delete_secret(path=name)
+    client_obj.delete_secret(path=name, key=key)
     click.echo("Done")
 
 
@@ -351,50 +365,70 @@ def delete(client_obj: client.VaultClientBase, name: str) -> None:
     "--path",
     multiple=True,
     required=True,
-    help="Folder or single item. Pass several times to load multiple values. You can use --path mypath=prefix if you want to change the generated names of the environment variables",
+    help="Folder or single item. Pass several times to load multiple values. You can use --path mypath=prefix or --path mypath:key=prefix if you want to change the generated names of the environment variables",
+)
+@click.option(
+    "-o",
+    "--omit-single-key/--no-omit-single-key",
+    is_flag=True,
+    default=False,
+    help="When the secret has only one key, don't use that key to build the name of the environment variable",
 )
 @click.argument("command", nargs=-1)
 @click.pass_obj
 @handle_errors()
 def env(
-    client_obj: client.VaultClientBase, path: Sequence[str], command: Sequence[str]
+    client_obj: client.VaultClientBase,
+    path: Sequence[str],
+    omit_single_key: bool,
+    command: Sequence[str],
 ) -> NoReturn:
     """
     Launch a command, loading secrets in environment.
 
     Strings are exported as-is, other types (including booleans, nulls, dicts, lists)
-    are exported as yaml (more specifically as json).
+    are exported JSON-encoded.
+
+    If the path ends with `:key` then only one key of the mapping is used and its name is the name of the key.
 
     VARIABLE NAMES
 
-    If the path is not a folder the prefix value is used as the name of the
-    variable.
-    e.g.: for a/b/key, `--path a/b/c=foo` gives `FOO=...`
+    By default the name is build upon the relative path to the parent of the given path (in parameter) and the name of the keys in the value mapping.
+    Let's say that we have stored the mapping `{'username': 'me', 'password': 'xxx'}` at path `a/b/c`
 
-    If the path is a folder, then the variable names will be generated as:
-    PREFIX + _ + relative path
-    e.g.: for a/b/key, `--path a/b=my` gives `MY_KEY=...`
+    Using `--path a/b` will inject the following environment variables: B_C_USERNAME and B_C_PASSWORD
+    Using `--path a/b/c` will inject the following environment variables: C_USERNAME and C_PASSWORD
+    Using `--path a/b/c:username` will only inject `USERNAME=me` in the environment.
 
-    The standard behavior when no prefix is set will use the relative path to
-    the parent of the given path as variable name
-    e.g.: for a/b/key, --path a/b gives `B_KEY=...`
-    e.g.: for a/b/key, --path a/b/key gives `KEY=...`
+    You can customize the variable names generation by appending `=SOME_PREFIX` to the path.
+    In this case the part corresponding to the base path is replace by your prefix.
+
+    Using `--path a/b=FOO` will inject the following environment variables: FOO_C_USERNAME and FOO_C_PASSWORD
+    Using `--path a/b/c=FOO` will inject the following environment variables: FOO_USERNAME and FOO_PASSWORD
+    Using `--path a/b/c:username=FOO` will inject `FOO=me` in the environment.
     """
     paths = list(path) or [""]
 
     env_secrets = {}
 
     for path in paths:
-        path, _, prefix = path.partition("=")
-        secrets = client_obj.get_secrets(path)
-        env_secrets.update(
-            {
-                environment.make_env_key(
-                    path=path, prefix=prefix, key=key
-                ): environment.make_env_value(value)
-                for key, value in secrets.items()
-            }
-        )
+        path_with_key, _, prefix = path.partition("=")
+        path, _, filter_key = path_with_key.partition(":")
+
+        if filter_key:
+            secret = client_obj.get_secret(path=path, key=filter_key)
+            env_updates = environment.get_envvars_for_secret(
+                key=filter_key, secret=secret, prefix=prefix
+            )
+        else:
+            secrets = client_obj.get_secrets(path=path, relative=True)
+            env_updates = environment.get_envvars_for_secrets(
+                path=path,
+                prefix=prefix,
+                secrets=secrets,
+                omit_single_key=omit_single_key,
+            )
+        env_secrets.update(env_updates)
 
     environ = os.environ.copy()
     environ.update(env_secrets)